requirements.txt
Django>=4.2
djangorestframework
drf-nested-routers
pytest
pytest-django
psycopg2-binary
flake8